Draft: Modernize C-style casts and disallow their use in ns-3

Closes #1315

See merge request nsnam/ns-3-dev!2731

+++ src/applications/model/onoff-application.cc
@@ -270,8 +270,8 @@ OnOffApplication::SendPacket()
         packet = Create<Packet>(m_pktSize);
     }
 
-    int actual = m_socket->Send(packet);
-    if ((unsigned)actual == m_pktSize)
+    uint32_t actual = m_socket->Send(packet);
+    if (actual == m_pktSize)
     {
         m_txTrace(packet);
         m_totBytes += m_pktSize;
