+++ src/stats/model/histogram.cc
@@ -69,7 +69,7 @@ Histogram::GetBinCount(uint32_t index) const
 void
 Histogram::AddValue(double value)
 {
-    auto index = (uint32_t)std::floor(value / m_binWidth);
+    auto index = static_cast<std::size_t>(std::floor(value / m_binWidth));
 
     // check if we need to resize the vector
     NS_LOG_DEBUG("AddValue: index=" << index << ", m_histogram.size()=" << m_histogram.size());
